requirements.txt
# AI
google-genai>=1.0.0

# X (Twitter) API
tweepy>=4.14.0

# Google Sheets
gspread>=6.0.0
google-auth>=2.0.0

# Environment
python-dotenv>=1.0.0

# HTTP
requests>=2.31.0

# Testing
pytest>=8.0.0
